Update SOURCEME.sh
As agreed,, removed Ed's new code that source's cpl-library's SOURCEME.sh has has not yet been done so.  This required changing directory to CPL_APP_OPENFOAM, which is not an expected behaviour and forced the user to change directory back.   This had not been documented so no further associated updates are required, i.e., everything works as advertised/expected.